création de l'objet Coup pour le minmax Action
